Move MultioTestEnvironment to a separate file
