+++ tests/multio/action/statistics/test_inverse_difference.cc
@@ -15,6 +15,7 @@
 #include "eckit/testing/Test.h"
 
 #include "testStatisticsUtils.h"
+#include "../../MultioTestEnvironment.h"
 
 
 namespace multio::test {
